Create federation package in tests and move federation related tests to it. Move FederationMember to federation package to avoid circular dependency. Rename int i to int version to use meaningful names
