+++ rskj-core/src/main/java/co/rsk/config/BridgeMainNetConstants.java
@@ -4,7 +4,7 @@ import co.rsk.bitcoinj.core.BtcECKey;
 import co.rsk.bitcoinj.core.Coin;
 import co.rsk.bitcoinj.core.NetworkParameters;
 import co.rsk.peg.AddressBasedAuthorizer;
-import co.rsk.peg.FederationMember;
+import co.rsk.peg.federation.FederationMember;
 import co.rsk.peg.federation.FederationFactory;
 import com.google.common.collect.Lists;
 import java.time.Instant;
