+++ rskj-core/src/main/java/co/rsk/peg/federation/Federation.java
@@ -23,7 +23,6 @@ import co.rsk.bitcoinj.core.BtcECKey;
 import co.rsk.bitcoinj.core.NetworkParameters;
 import co.rsk.bitcoinj.script.Script;
 import co.rsk.bitcoinj.script.ScriptBuilder;
-import co.rsk.peg.FederationMember;
 
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
