+++ rskj-core/src/main/java/co/rsk/peg/federation/FederationMember.java
@@ -16,7 +16,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package co.rsk.peg;
+package co.rsk.peg.federation;
 
 import co.rsk.bitcoinj.core.BtcECKey;
 import com.google.common.primitives.UnsignedBytes;
