+++ rskj-core/src/main/java/co/rsk/peg/federation/StandardMultisigFederation.java
@@ -21,7 +21,6 @@ package co.rsk.peg.federation;
 import co.rsk.bitcoinj.core.NetworkParameters;
 import co.rsk.bitcoinj.script.Script;
 import co.rsk.bitcoinj.script.ScriptBuilder;
-import co.rsk.peg.FederationMember;
 import co.rsk.peg.bitcoin.ScriptValidations;
 
 import java.time.Instant;
