+++ rskj-core/src/main/java/co/rsk/remasc/RemascFederationProvider.java
@@ -20,7 +20,7 @@ package co.rsk.remasc;
 import co.rsk.config.BridgeConstants;
 import co.rsk.core.RskAddress;
 import co.rsk.peg.BridgeStorageProvider;
-import co.rsk.peg.FederationSupport;
+import co.rsk.peg.federation.FederationSupport;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
 import org.ethereum.core.Block;
 import org.ethereum.core.Repository;
