+++ rskj-core/src/test/java/co/rsk/peg/BridgeUtilsLegacyTest.java
@@ -8,6 +8,7 @@ import java.time.Instant;
 import java.util.List;
 import co.rsk.peg.federation.Federation;
 import co.rsk.peg.federation.FederationFactory;
+import co.rsk.peg.federation.FederationMember;
 import org.bouncycastle.util.encoders.Hex;
 import org.ethereum.config.blockchain.upgrades.ActivationConfig;
 import org.ethereum.config.blockchain.upgrades.ConsensusRule;
